* gamewin.cc, exult.cc, keys.cc: fixed bug #3011719  avatar can do stuff   while passed out. The player will be unable to do almost anything (similar   to the originals) if the avatar is asleep, unconscious, paralyzed or dead.   As a rule, cheat and mapedit actions are still allowed in these cases, and   hackmove allows dragging even in these conditions. Also, purely interface   actions will also work -- like esc, scrolling the map with the mouse wheel   and so on. * configure.ac: AC_PATH_XTRA is actually causing linking problems in Ubuntu   Karmic, so it has been removed (again). * Others: Eliminated a few warnings.

+++ exult/trunk/audio/Midi.cc
@@ -160,7 +160,7 @@ void	MyMidiPlayer::start_music(int num,bool repeat,std::string flex)
 
 	// Try in patch dir first.
 	string pflex("<PATCH>/");
-	int prefix_len = 0;
+	size_t prefix_len = 0;
 	if (flex[0] == '<')
 	{
 		prefix_len = flex.find(">/");
